Change error messaging to allow for easier debugging.
