+++ Test/TestCase.h
@@ -1,4 +1,12 @@
+// Licensed under the MIT License.
+// Copyright David LeBlanc - dcl@dleblanc.net
+
 #pragma once
+
+#include <iostream>
+#include <ios>
+#include <iomanip>
+#include <sstream>
 
 enum OpType
 {
@@ -76,3 +84,46 @@ private:
 	size_t current;
 };
 
+#if !defined(COUNTOF)
+# if defined(_countof)
+#  define COUNTOF(x) _countof(x)
+# else
+#  define COUNTOF(x) (sizeof(x)/sizeof((x)[0]))
+# endif
+#endif
+
+template <typename T>
+std::string to_hex(T t)
+{
+	std::ostringstream ostm;
+	ostm << "0x" << std::setfill('0') << std::hex << std::setw(sizeof(t) <= 4 ? 8 : 16) << t;
+	return ostm.str();
+}
+
+template <>
+inline std::string to_hex< uint8_t >(uint8_t t)
+{
+	std::ostringstream ostm;
+	ostm << "0x" << std::setfill('0') << std::hex << std::setw(2) << static_cast<uint16_t>(t);
+	return ostm.str();
+}
+
+template <>
+inline std::string to_hex< int8_t >(int8_t t)
+{
+	std::ostringstream ostm;
+	ostm << "0x" << std::setfill('0') << std::hex << std::setw(2) << static_cast<uint16_t>(t);
+	return ostm.str();
+}
+
+template <typename T, typename U>
+void err_msg(const std::string& msg, T t, U u, bool expected)
+{
+	std::cerr << msg << to_hex(t) << ", " << to_hex(u) << ", expected = " << expected << std::endl;
+}
+
+template <typename T>
+void err_msg(const std::string& msg, T t, bool expected)
+{
+	std::cerr << msg << to_hex(t) << ", expected = " << expected << std::endl;
+}
